backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
django-cors-headers>=4.3
Pillow>=10.3
python-dotenv>=1.0
gunicorn>=22.0
